use decorator syntax instead. add comments
